# api/api.yaml
swagger: '2.0'
info:
  title: Bubbles API
  description: Mrzim svoj život.
  version: 1.0.0
produces:
  - application/json
paths:
...
  '/content/comments/{content_id}':
    get:
      tags:
        - Content
      summary: Vraća sve komentare za dani content_id
      parameters:
        - name: content_id
          in: path
          required: true
          type: integer
          format: int32
      responses:
        '200':
          description: Komentari
          schema:
            type: object
            properties:
              comments:
                type: array
                items:
                  $ref: '#/definitions/ExtendedComment'
  '/content/comment/{content_id}':
    post:
      tags:
        - Content
      summary: Posta komentar na content sa content_id-om
      parameters:
        - name: content_id
          in: path
          required: true
          type: integer
          format: int32
      responses:
        '200':
          description: Kod uspjesnog komentiranja
  '/content/like/{id}':
    post:
      tags:
        - Content
      summary: Lajka/unlajka(toggle) content sa idem id
      parameters:
        - name: id
          in: path
          required: true
          type: integer
          format: int32
      responses:
        '200':
          description: Uspješno lajkanje
  '/content/likes/{id}':
    get:
      tags:
        - Content
      summary: Dohvaća kolekciju korisnika koji su lajkali content sa id-em id
      parameters:
        - name: id
          in: path
          required: true
          type: integer
          format: int32
      responses:
        '200':
          description: Uspješno dohvaćanje
          schema:
            type: object
            properties:
              users:
                type: array
                items:
                  type: object
                  properties:
                    first_name:
                      type: string
                    last_name:
                      type: string
                    middle_name:
                      type: string
                    username:
                      type: string
  /bubble/create:
    post:
      tags:
        - Bubble
      summary: Stvara novi bubble za aktualnog vlasnika session-a
      parameters:
        - name: title
          in: formData
          required: true
          type: string
        - name: description
          in: formData
          required: false
          type: string
      responses:
        '200':
          description: Kod uspješnog spremanja u bazu.
        default:
          description: Ako dođe do greške tokom spremanja.
          schema:
            $ref: '#/definitions/ValidationError'
  '/bubble/edit/{id}':
    post:
      tags:
        - Bubble
      summary: Omogućava promjenu naslova i opisa postojećeg bubble-a.
      parameters:
        - name: id
          in: path
          required: true
          type: integer
        - name: title
          in: formData
          required:  true
          type: string
        - name: description
          in: formData
          required: false
          type: string
      responses:
        '200':
          description: Uspješna promjena podataka.
          schema:
            $ref: '#/definitions/Bubble'
        '400':
          description: Nedozvoljeni podaci.
          schema:
            $ref: '#/definitions/ValidationError'
        '404':
          description: Pokušaj promjene neposojećeg bubble-a.
          schema:
            $ref: '#/definitions/Error'
        '403':
          description: Pokušaj promjene tuđeg bubble-a.
          schema:
            $ref: '#/definitions/Error'
...
definitions:
...
  ExtendedComment:
    type: object
    properties:
      id:
        type: integer
        format: int32
      content_id:
        type: integer
        format: int32
      user_id:
        type: integer
        format: int32
      created_at:
        type: string
      updated_at:
        type: string
      comment:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      middle_name:
        type: string
      username:
        type: string
  Bubble:
    type: object
    properties:
      user_id:
        type: integer
      id:
        type: integer
      bubble_type_id:
        type: integer
      created_at:
        type: string
      updated_at:
        type: string
      title:
        type: string
      description:
        type: string
...
  Error:
    type: object
    properties:
      error:
        type: string
  ValidationError:
    type: object
    properties:
      errors:
        type: array
        items:
          type: string